Scalebuttons Functioning!
I got the scale buttons functioning properly, added a scale readout,
added an array of scale values (insted of just an integer counting
variable).  The array happens to contain integers from 1-10 at the
moment, but this is because we don't know how NE's 0-1 scale matches up
to the real world in micro-volts, so when we know that info we can
update the scale-factor array to correspond to 40-400 micro-volts or
something.

I also changed the color palette of the project some and added a quad as
a background so we could begin moving away from that generic blue.