Recording and LSL Markers Working!
We are managing to record .easy files from the Unity GUI.  The recording
begins automatically (probably should make this a button).  We have not
recorded .info files which contain such info as time, subject name,
num_electrodes, etc.

I also got LSL markers working with a button in the GUI.  The markers
show up in the debug_log and also in the .easy file.

An error message showed up but does not seem to correspond to any issue
with program functionality.  It is at line 68 (is it?) in
LineRendererScript.cs and is of type "something here does not refer to a
real object"  I dunno.